Move dither to imageproperty
