+++ meerk40t/gui/propertypanels/rasterwizardpanels.py
@@ -971,94 +971,6 @@ class GammaPanel(wx.Panel):
         pass
 
 
-class DitherPanel(wx.Panel):
-    name = _("Dither")
-    priority = 35
-
-    @staticmethod
-    def accepts(node):
-        if node.type != "elem image":
-            return False
-        for n in node.operations:
-            if n.get("name") == "dither":
-                return True
-        return False
-
-    def __init__(self, *args, context=None, node=None, **kwds):
-        kwds["style"] = kwds.get("style", 0) | wx.TAB_TRAVERSAL
-        wx.Panel.__init__(self, *args, **kwds)
-        self.context = context
-        self.node = node
-
-        self.check_enable_dither = wx.CheckBox(self, wx.ID_ANY, _("Enable"))
-        self.choices = [
-            "Floyd-Steinberg",
-            "Atkinson",
-            "Jarvis-Judice-Ninke",
-            "Stucki",
-            "Burkes",
-            "Sierra3",
-            "Sierra2",
-            "Sierra-2-4a",
-        ]
-        self.combo_dither = wx.ComboBox(
-            self,
-            wx.ID_ANY,
-            choices=self.choices,
-            style=wx.CB_DROPDOWN,
-        )
-
-        self.__set_properties()
-        self.__do_layout()
-
-        self.Bind(
-            wx.EVT_CHECKBOX, self.on_check_enable_dither, self.check_enable_dither
-        )
-        self.Bind(wx.EVT_COMBOBOX, self.on_combo_dither_type, self.combo_dither)
-        self.Bind(wx.EVT_TEXT_ENTER, self.on_combo_dither_type, self.combo_dither)
-
-        op = None
-        for n in node.operations:
-            if n.get("name") == "dither":
-                op = n
-        if op is None:
-            raise ValueError
-        self.op = op
-        self.original_op = deepcopy(op)
-        self.check_enable_dither.SetValue(op["enable"])
-        self.combo_dither.SetSelection(self.choices.index(self.op["type"]))
-
-    def __set_properties(self):
-        # begin wxGlade: DitherPanel.__set_properties
-        self.check_enable_dither.SetToolTip(_("Enable Dither"))
-        self.check_enable_dither.SetValue(1)
-        self.combo_dither.SetToolTip(_("Select dither algorithm to use"))
-        self.combo_dither.SetSelection(0)
-        # end wxGlade
-
-    def __do_layout(self):
-        # begin wxGlade: DitherPanel.__do_layout
-        sizer_dither_main = wx.StaticBoxSizer(
-            wx.StaticBox(self, wx.ID_ANY, _("Dither")), wx.VERTICAL
-        )
-        sizer_dither_main.Add(self.check_enable_dither, 0, 0, 0)
-        sizer_dither_main.Add(self.combo_dither, 0, 0, 0)
-        self.SetSizer(sizer_dither_main)
-        sizer_dither_main.Fit(self)
-        self.Layout()
-        # end wxGlade
-
-    def on_check_enable_dither(
-        self, event=None
-    ):  # wxGlade: RasterWizard.<event_handler>
-        self.op["enable"] = self.check_enable_dither.GetValue()
-        self.node.update(self.context)
-
-    def on_combo_dither_type(self, event=None):  # wxGlade: RasterWizard.<event_handler>
-        self.op["type"] = self.choices[self.combo_dither.GetSelection()]
-        self.node.update(self.context)
-
-
 class EdgePanel(wx.Panel):
     name = _("Edge Enhance")
     priority = 40
